Refactoring of paths.rs
