+++ bitloops/src/utils/paths/constants.rs
@@ -1,0 +1,24 @@
+pub const BITLOOPS_DIR: &str = ".bitloops";
+pub const BITLOOPS_TMP_DIR: &str = ".bitloops/tmp";
+pub const BITLOOPS_METADATA_DIR: &str = ".bitloops/metadata";
+pub const BITLOOPS_STORES_DIR: &str = ".bitloops/stores";
+pub const BITLOOPS_RELATIONAL_STORE_DIR: &str = ".bitloops/stores/relational";
+pub const BITLOOPS_EVENT_STORE_DIR: &str = ".bitloops/stores/event";
+pub const BITLOOPS_BLOB_STORE_DIR: &str = ".bitloops/stores/blob";
+pub const BITLOOPS_EMBEDDINGS_DIR: &str = ".bitloops/embeddings";
+pub const BITLOOPS_EMBEDDING_MODELS_DIR: &str = ".bitloops/embeddings/models";
+pub const RELATIONAL_DB_FILE_NAME: &str = "relational.db";
+pub const EVENTS_DB_FILE_NAME: &str = "events.duckdb";
+
+pub const CONTEXT_FILE_NAME: &str = "context.md";
+pub const PROMPT_FILE_NAME: &str = "prompt.txt";
+pub const SUMMARY_FILE_NAME: &str = "summary.txt";
+pub const TRANSCRIPT_FILE_NAME: &str = "full.jsonl";
+pub const TRANSCRIPT_FILE_NAME_LEGACY: &str = "full.log";
+pub const METADATA_FILE_NAME: &str = "metadata.json";
+pub const CHECKPOINT_FILE_NAME: &str = "checkpoint.json";
+pub const CONTENT_HASH_FILE_NAME: &str = "content_hash.txt";
+pub const EXPORT_DATA_FILE_NAME: &str = "export.json";
+pub const SETTINGS_FILE_NAME: &str = "settings.json";
+
+pub const METADATA_BRANCH_NAME: &str = "bitloops/checkpoints/v1";
